+++ src/builtin_functions/round.cpp
@@ -18,12 +18,12 @@ namespace round_function {
 
     void evaluate_floor(Term* caller)
     {
-        as_int(caller) = std::floor(float_input(caller, 0));
+        as_int(caller) = (int) std::floor(float_input(caller, 0));
     }
 
     void evaluate_ceil(Term* caller)
     {
-        as_int(caller) = std::ceil(float_input(caller, 0));
+        as_int(caller) = (int) std::ceil(float_input(caller, 0));
     }
 
     void setup(Branch& kernel)
